LIN-13 refactor: doropdown 종류 분리 및 ref 활용 재구성 / 상수 분리
