+++ src/constants/ProductsConst.tsx
@@ -1,0 +1,32 @@
+interface Category {
+  id: number;
+  name: string;
+}
+
+export const CATEGORY: Category[] = [
+  { id: 1, name: '카테고리1' },
+  { id: 2, name: '카테고리2' },
+  { id: 3, name: '카테고리3' },
+  { id: 4, name: '카테고리4' },
+  { id: 5, name: '카테고리5' },
+  { id: 6, name: '카테고리6' },
+  { id: 7, name: '카테고리7' },
+];
+
+interface Sort {
+  value: string;
+  name: string;
+}
+
+export const SORT_OPTION_PRODUCTS: Sort[] = [
+  { value: 'recent', name: '조회순' },
+  { value: 'ratingDesc', name: '별점 높은순' },
+  { value: 'ratingAsc', name: '별점 낮은순' },
+];
+
+export const SORT_OPTION_REVIEWS: Sort[] = [
+  { value: 'recent', name: '조회순' },
+  { value: 'ratingDesc', name: '별점 높은순' },
+  { value: 'ratingAsc', name: '별점 낮은순' },
+  { value: 'likeCount', name: '좋아요순' },
+];
